+++ src/frontend/src/pages/NavBar.vue
